+++ evalml/utils/gen_utils.py
@@ -216,6 +216,7 @@ _not_used_in_automl = {
     "SVMClassifier",
     "SVMRegressor",
     "LinearRegressor",
+    "VARMAXRegressor",
     "VowpalWabbitBinaryClassifier",
     "VowpalWabbitMulticlassClassifier",
     "VowpalWabbitRegressor",