+++ docs/source/release_notes.rst
@@ -2,12 +2,29 @@ Release Notes
 -------------
 **Future Releases**
     * Enhancements
+    * Fixes
+    * Changes
+    * Documentation Changes
+    * Testing Changes
+
+.. warning::
+
+    **Breaking Changes**
+
+
+**v0.79.0 Aug. 11, 2023**
+    * Enhancements
         * Updated regression metrics to handle multioutput dataframes as well as single output series :pr:`4233`
         * Added baseline regressor for multiseries time series problems :pr:`4246`
         * Added stacking and unstacking utility functions to work with multiseries data :pr:`4250`
+        * Added multiseries regression pipeline class :pr:`4256`
+        * Added multiseries VARMAX regressor :pr:`4238`
         * Extend STLDecomposer to Support Multiseries :pr:`4253`
     * Fixes
         * Added support for pandas 2 :pr:`4216`
+        * Fixed bug where time series pipelines would fail due to MASE needing `y_train` when scoring :pr:`4258`
+        * Update s3 bucket for docs image :pr:`4260`
+        * Fix deps checker including any package with post in the name :pr:`4268`
     * Changes
         * Unpinned sktime version :pr:`4214`
         * Bumped minimum lightgbm version to 4.0.0 for nullable type handling :pr:`4237`